Implement better email validation support
These changes allow for multiple email validators to be added when checking for valid emails. This is a continuation of the previous PR: https://github.com/laravel/framework/pull/26503

Basically this allows for two things:

- Make use of multiple email validators provided by the egulias/email-validator package
- Use the previous (and much requested) filter_var validation

By default nothing's breaking because it'll still use the RFC validator to when no validators are passed to the email validation rule. But you can opt to include different ones or multiple ones:

    'email' => 'email:rfc,dns'

Or opt for the pre-5.8 behavior:

    'email' => 'email:filter'

Which will use `filter_var` to validate the email address. This should give people a little more flexibility when doing email validation.